Fix a bug in a package checksum calculation
Package checksum calculation wasn't correct because a relative file path
from the download directory was given to the `updateSha1Checksum`
procedure and it being not able to find the files was exiting early
instead of opening and reading them. Now the procedure is changed to get
both relative and full path to the files. The first is used for
accumulating the file name into the checksum and the second for the
actual opening and reading the file.

Related to nim-lang/nimble#127